Expose API for app categories
