+++ app-tracking-protection/vpn-impl/src/test/java/com/duckduckgo/mobile/android/vpn/ui/tracker_activity/DeviceShieldActivityFeedViewModelTest.kt
@@ -24,9 +24,9 @@ import com.duckduckgo.app.CoroutineTestRule
 import com.duckduckgo.app.global.formatters.time.DatabaseDateFormatter
 import com.duckduckgo.app.global.formatters.time.RealTimeDiffFormatter
 import com.duckduckgo.mobile.android.vpn.AppTpVpnFeature
-import com.duckduckgo.mobile.android.vpn.apps.AppCategory
 import com.duckduckgo.mobile.android.vpn.apps.TrackingProtectionAppInfo
 import com.duckduckgo.mobile.android.vpn.apps.TrackingProtectionAppsRepository
+import com.duckduckgo.mobile.android.vpn.exclusion.AppCategory
 import com.duckduckgo.mobile.android.vpn.model.TrackingApp
 import com.duckduckgo.mobile.android.vpn.model.VpnTracker
 import com.duckduckgo.mobile.android.vpn.state.VpnStateMonitor
@@ -376,7 +376,6 @@ private val config = DeviceShieldActivityFeedFragment.ActivityFeedConfig(
 private val app = TrackingProtectionAppInfo(
     packageName = "com.package.name",
     name = "App",
-    type = "None",
     category = AppCategory.Undefined,
     isExcluded = false,
     knownProblem = TrackingProtectionAppInfo.NO_ISSUES,
@@ -386,7 +385,6 @@ private val app = TrackingProtectionAppInfo(
 private val excludedApp = TrackingProtectionAppInfo(
     packageName = "com.package.name",
     name = "App",
-    type = "None",
     category = AppCategory.Undefined,
     isExcluded = true,
     knownProblem = TrackingProtectionAppInfo.NO_ISSUES,
