+++ network-protection/network-protection-impl/src/main/java/com/duckduckgo/networkprotection/impl/exclusion/TrackingProtectionAppInfo.kt
@@ -14,50 +14,13 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.mobile.android.vpn.apps
+package com.duckduckgo.networkprotection.impl.exclusion
 
-sealed class AppCategory {
-    object Undefined : AppCategory() {
-        override fun toString() = "CATEGORY_UNDEFINED"
-    }
-
-    object Game : AppCategory() {
-        override fun toString() = "CATEGORY_GAME"
-    }
-
-    object Audio : AppCategory() {
-        override fun toString() = "CATEGORY_AUDIO"
-    }
-
-    object Video : AppCategory() {
-        override fun toString() = "CATEGORY_VIDEO"
-    }
-
-    object Image : AppCategory() {
-        override fun toString() = "CATEGORY_IMAGE"
-    }
-
-    object Social : AppCategory() {
-        override fun toString() = "CATEGORY_SOCIAL"
-    }
-
-    object News : AppCategory() {
-        override fun toString() = "CATEGORY_NEWS"
-    }
-
-    object Maps : AppCategory() {
-        override fun toString() = "CATEGORY_MAPS"
-    }
-
-    object Productivity : AppCategory() {
-        override fun toString() = "CATEGORY_PRODUCTIVITY"
-    }
-}
+import com.duckduckgo.mobile.android.vpn.exclusion.AppCategory
 
 data class TrackingProtectionAppInfo(
     val packageName: String,
     val name: String,
-    val type: String? = null,
     val category: AppCategory = AppCategory.Undefined,
     val isExcluded: Boolean = false,
     val knownProblem: Int,
